requirements.txt
lxml>=3.4.4
shapely>=1.5.12
pycollada>=0.4.1
pyproj>=1.9.4
